Dynamic Page titles Fixes #46
