+++ frontend/public/index.html
@@ -7,7 +7,7 @@
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 	<script src="https://kit.fontawesome.com/3ebd1745ed.js" crossorigin="anonymous"></script>
 	<link rel="icon" type="image/x-icon" href="./images/icon.ico">
-    <title>Best Notetaking App - Organize Your Notes with Evernote Clone</title>
+    <title>Second Best Notetaking App - Organize Your Notes with Evernote Clone</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
