+++ frontend/src/components/Landing/valueProp.js
@@ -1,6 +1,11 @@
+import { useEffect } from "react";
 import { Link } from "react-router-dom";
 
 export default function ValueProp() {
+	useEffect(() => {
+		document.title = "Second Best Notetaking App - Organize Your Notes with Evernote Clone";
+	}, [])
+
 	return (
 		<div id="value-prop-div">
 			<h1 id="title-h1">Tame your work, organize your life</h1>
